refactor: PersonalCharacteristics  -> SurveyDetail code cleanup currently unused classes
