+++ lib/views/survey/shared/lorem_ipsum.dart
